+++ src/api/block_builder/input_selection/mod.rs
@@ -4,6 +4,7 @@
 //! Input selection for transactions
 
 mod automatic;
+mod helpers;
 mod manual;
 mod native_token_helpers;
 mod output_data;
@@ -12,17 +13,13 @@ pub mod types;
 use std::{collections::HashSet, ops::Deref};
 
 use bee_block::{
-    address::{Address, AliasAddress, Ed25519Address, NftAddress},
+    address::{Address, AliasAddress, NftAddress},
     input::INPUT_COUNT_MAX,
     output::{
-        unlock_condition::{
-            AddressUnlockCondition, GovernorAddressUnlockCondition, ImmutableAliasAddressUnlockCondition,
-            StateControllerAddressUnlockCondition, StorageDepositReturnUnlockCondition,
-        },
-        AliasOutputBuilder, BasicOutputBuilder, FoundryOutputBuilder, NativeTokens, NftOutputBuilder, Output, Rent,
-        RentStructure, UnlockCondition, OUTPUT_COUNT_MAX,
+        AliasOutputBuilder, FoundryOutputBuilder, NftOutputBuilder, Output, Rent, RentStructure, OUTPUT_COUNT_MAX,
     },
 };
+pub use helpers::minimum_storage_deposit;
 use packable::bounded::TryIntoBoundedU16Error;
 
 use self::{
@@ -32,16 +29,20 @@ use self::{
     types::SelectedTransactionData,
 };
 use crate::{
-    api::input_selection::{remainder::get_storage_deposit_return_outputs, types::AccumulatedOutputAmounts},
+    api::input_selection::{
+        helpers::{output_contains_address, sdr_not_expired, sort_input_signing_data},
+        remainder::get_storage_deposit_return_outputs,
+        types::AccumulatedOutputAmounts,
+    },
     secret::types::InputSigningData,
     Error, Result,
 };
 
 /// Select inputs from provided inputs([InputSigningData]) for provided [Output]s, validate amounts and create remainder
 /// output if necessary. Also checks for alias, foundry and nft outputs that there previous output exist in the inputs,
-/// when required. Careful with setting `allow_burning` to `true`, native tokens can get easily burned by accident.
-/// Provided alias, foundry and nft outputs will be used first as inputs and therefore destroyed if not existent in the
-/// output
+/// when required. Careful with setting `allow_burning` to `true`, native tokens, nfts or alias outputs can get easily
+/// burned by accident. Without burning, alias, foundry and nft outputs will be created on the output side, if not
+/// already present.
 pub fn try_select_inputs(
     mut inputs: Vec<InputSigningData>,
     mut outputs: Vec<Output>,
@@ -153,36 +154,7 @@ pub fn try_select_inputs(
                         let nft_address = Address::Nft(NftAddress::new(nft_id));
                         let nft_required_in_unlock_condition = outputs.iter().any(|output| {
                             // check if alias address is in unlock condition
-                            if let Some(unlock_conditions) = output.unlock_conditions() {
-                                if let Some(UnlockCondition::Address(address_unlock_condition)) =
-                                    unlock_conditions.get(AddressUnlockCondition::KIND)
-                                {
-                                    if &nft_address
-                                        == unlock_conditions
-                                            .locked_address(address_unlock_condition.address(), current_time)
-                                    {
-                                        return true;
-                                    }
-                                }
-                                if let Some(UnlockCondition::StateControllerAddress(
-                                    state_controller_unlock_condition,
-                                )) = unlock_conditions.get(StateControllerAddressUnlockCondition::KIND)
-                                {
-                                    if &nft_address == state_controller_unlock_condition.address() {
-                                        return true;
-                                    }
-                                }
-                                if let Some(UnlockCondition::GovernorAddress(governor_controller_unlock_condition)) =
-                                    unlock_conditions.get(GovernorAddressUnlockCondition::KIND)
-                                {
-                                    if &nft_address == governor_controller_unlock_condition.address() {
-                                        return true;
-                                    }
-                                }
-                                false
-                            } else {
-                                false
-                            }
+                            output_contains_address(output, &nft_address, current_time)
                         });
 
                         // Don't add if it doesn't give us any amount or native tokens
@@ -215,44 +187,7 @@ pub fn try_select_inputs(
                         let alias_address = Address::Alias(AliasAddress::new(alias_id));
                         let alias_required_in_unlock_condition = outputs.iter().any(|output| {
                             // check if alias address is in unlock condition
-                            if let Some(unlock_conditions) = output.unlock_conditions() {
-                                if let Some(UnlockCondition::Address(address_unlock_condition)) =
-                                    unlock_conditions.get(AddressUnlockCondition::KIND)
-                                {
-                                    if &alias_address
-                                        == unlock_conditions
-                                            .locked_address(address_unlock_condition.address(), current_time)
-                                    {
-                                        return true;
-                                    }
-                                }
-                                if let Some(UnlockCondition::StateControllerAddress(
-                                    state_controller_unlock_condition,
-                                )) = unlock_conditions.get(StateControllerAddressUnlockCondition::KIND)
-                                {
-                                    if &alias_address == state_controller_unlock_condition.address() {
-                                        return true;
-                                    }
-                                }
-                                if let Some(UnlockCondition::GovernorAddress(governor_controller_unlock_condition)) =
-                                    unlock_conditions.get(GovernorAddressUnlockCondition::KIND)
-                                {
-                                    if &alias_address == governor_controller_unlock_condition.address() {
-                                        return true;
-                                    }
-                                }
-                                if let Some(UnlockCondition::ImmutableAliasAddress(
-                                    immutable_alias_address_unlock_condition,
-                                )) = unlock_conditions.get(ImmutableAliasAddressUnlockCondition::KIND)
-                                {
-                                    if &alias_address == immutable_alias_address_unlock_condition.address() {
-                                        return true;
-                                    }
-                                }
-                                false
-                            } else {
-                                false
-                            }
+                            output_contains_address(output, &alias_address, current_time)
                         });
 
                         // Don't add if it doesn't give us any amount or native tokens
@@ -564,134 +499,4 @@ pub fn try_select_inputs(
         outputs,
         remainder: remainder_data,
     })
-}
-
-// Inputs need to be sorted because the reference unlock conditions can only reference a lower index
-fn sort_input_signing_data(inputs: Vec<InputSigningData>) -> crate::Result<Vec<InputSigningData>> {
-    // filter for ed25519 address first, safe to unwrap since we encoded it before
-    let mut sorted_inputs = inputs
-        .clone()
-        .into_iter()
-        .filter(|input| Address::try_from_bech32(&input.bech32_address).unwrap().1.kind() == Ed25519Address::KIND)
-        .collect::<Vec<InputSigningData>>();
-
-    for input in &inputs {
-        // Don't add outputs duplicated
-        if sorted_inputs
-            .iter()
-            .any(|i| i.output_id().unwrap() == input.output_id().unwrap())
-        {
-            continue;
-        }
-
-        match sorted_inputs.iter().position(|input_signing_data| {
-            match Address::try_from_bech32(&input.bech32_address) {
-                Ok((_, unlock_address)) => match unlock_address {
-                    Address::Alias(unlock_address) => {
-                        if let Output::Alias(alias_output) = &input_signing_data.output {
-                            *unlock_address.alias_id()
-                                == alias_output
-                                    .alias_id()
-                                    .or_from_output_id(input_signing_data.output_id().expect("Invalid output id"))
-                        } else {
-                            false
-                        }
-                    }
-                    Address::Nft(unlock_address) => {
-                        if let Output::Nft(nft_output) = &input_signing_data.output {
-                            *unlock_address.nft_id()
-                                == nft_output
-                                    .nft_id()
-                                    .or_from_output_id(input_signing_data.output_id().expect("Invalid output id"))
-                        } else {
-                            false
-                        }
-                    }
-                    _ => false,
-                },
-                _ => false,
-            }
-        }) {
-            Some(position) => {
-                // Insert after the output we need
-                sorted_inputs.insert(position + 1, input.clone());
-            }
-            None => {
-                // insert before address
-                let alias_or_nft_address = match &input.output {
-                    Output::Alias(alias_output) => Some(Address::Alias(AliasAddress::new(
-                        alias_output
-                            .alias_id()
-                            .or_from_output_id(input.output_id().expect("Invalid output id")),
-                    ))),
-                    Output::Nft(nft_output) => Some(Address::Nft(NftAddress::new(
-                        nft_output
-                            .nft_id()
-                            .or_from_output_id(input.output_id().expect("Invalid output id")),
-                    ))),
-                    _ => None,
-                };
-
-                if let Some(alias_or_nft_address) = alias_or_nft_address {
-                    // Check for existing outputs for this address, and insert before
-                    match sorted_inputs.iter().position(|input_signing_data| {
-                        Address::try_from_bech32(&input_signing_data.bech32_address)
-                            .expect("Safe to unwrap, we encoded it before")
-                            .1
-                            == alias_or_nft_address
-                    }) {
-                        Some(position) => {
-                            // Insert before the output with this address required for unlocking
-                            sorted_inputs.insert(position, input.clone());
-                        }
-                        // just push output
-                        None => sorted_inputs.push(input.clone()),
-                    }
-                } else {
-                    // just push basic or foundry output
-                    sorted_inputs.push(input.clone());
-                }
-            }
-        }
-    }
-
-    Ok(sorted_inputs)
-}
-
-/// Computes the minimum amount that an output needs to have, when native tokens are sent with [AddressUnlockCondition].
-pub fn minimum_storage_deposit(
-    config: &RentStructure,
-    address: &Address,
-    native_tokens: &Option<NativeTokens>,
-) -> Result<u64> {
-    let address_condition = UnlockCondition::Address(AddressUnlockCondition::new(*address));
-    let mut basic_output_builder = BasicOutputBuilder::new_with_amount(OutputAmount::MIN)?;
-    if let Some(native_tokens) = native_tokens {
-        basic_output_builder = basic_output_builder.with_native_tokens(native_tokens.clone());
-    }
-    let basic_output = basic_output_builder
-        .add_unlock_condition(address_condition)
-        .finish_output()?;
-
-    Ok(basic_output.rent_cost(config))
-}
-
-/// Get the `StorageDepositReturnUnlockCondition`, if not expired
-pub(crate) fn sdr_not_expired(output: &Output, current_time: u32) -> Option<&StorageDepositReturnUnlockCondition> {
-    if let Some(unlock_conditions) = output.unlock_conditions() {
-        if let Some(sdr) = unlock_conditions.storage_deposit_return() {
-            let expired = if let Some(expiration) = unlock_conditions.expiration() {
-                current_time >= expiration.timestamp()
-            } else {
-                false
-            };
-
-            // We only have to send the storage deposit return back if the output is not expired
-            if !expired { Some(sdr) } else { None }
-        } else {
-            None
-        }
-    } else {
-        None
-    }
 }
